java - using decimal for monetary values

+++ java/yapay-gateway-client/src/main/java/br/com/yapay/gateway/model/CheckoutData.java
@@ -1,5 +1,6 @@
 package br.com.yapay.gateway.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -69,7 +70,7 @@ public class CheckoutData {
 		this.multipleBillDataList = multipleBillDataList;
 	}
 
-	public void addMultipleBillData(Long value, LocalDate dueDate) {
+	public void addMultipleBillData(BigDecimal value, LocalDate dueDate) {
 		addMultipleBillData(new TransactionCheckoutMultipleBillData(value, dueDate));
 	}
